[TASK] Saving of LastSearches should be done with LastSearchesService instead of LastSearches ResultSetModifier
* Moved logic save logic from modifier into service and use the service

Fixes: #338
