+++ Classes/ResultsetModifier/LastSearches.php
@@ -26,6 +26,7 @@ namespace ApacheSolrForTypo3\Solr\ResultsetModifier;
 
 use ApacheSolrForTypo3\Solr\Plugin\Results\ResultsCommand;
 use ApacheSolrForTypo3\Solr\System\Configuration\TypoScriptConfiguration;
+use TYPO3\CMS\Core\Utility\GeneralUtility;
 
 /**
  * Logs the keywords from the query into the user's session or the database -
@@ -37,14 +38,6 @@ use ApacheSolrForTypo3\Solr\System\Configuration\TypoScriptConfiguration;
  */
 class LastSearches implements ResultSetModifier
 {
-
-    protected $prefix = 'tx_solr';
-
-    /**
-     * @var TypoScriptConfiguration
-     */
-    protected $configuration;
-
 
     /**
      * Does not actually modify the result set, but tracks the search keywords.
@@ -59,7 +52,6 @@ class LastSearches implements ResultSetModifier
         ResultsCommand $resultCommand,
         array $resultSet
     ) {
-        $this->configuration = $resultCommand->getParentPlugin()->getConfiguration();
         $keywords = $resultCommand->getParentPlugin()->getSearchResultSetService()->getSearch()->getQuery()->getKeywordsCleaned();
 
         $keywords = trim($keywords);
@@ -67,101 +59,16 @@ class LastSearches implements ResultSetModifier
             return $resultSet;
         }
 
-        switch ($this->configuration->getSearchLastSearchesMode()) {
-            case 'user':
-                $this->storeKeywordsToSession($keywords);
-                break;
-            case 'global':
-                $this->storeKeywordsToDatabase($keywords);
-                break;
-            default:
-                throw new \UnexpectedValueException(
-                    'Unknown mode for plugin.tx_solr.search.lastSearches.mode, valid modes are "user" or "global".',
-                    1342456570
-                );
-        }
+        $configuration = $resultCommand->getParentPlugin()->getConfiguration();
+
+            /** @var $lastSearchesService \ApacheSolrForTypo3\Solr\Domain\Search\LastSearches\LastSearchesService */
+        $lastSearchesService = GeneralUtility::makeInstance('ApacheSolrForTypo3\Solr\Domain\Search\LastSearches\LastSearchesService',
+            $configuration,
+            $GLOBALS['TSFE'],
+            $GLOBALS['TYPO3_DB']);
+
+        $lastSearchesService->addToLastSearches($keywords);
 
         return $resultSet;
-    }
-
-    /**
-     * Stores the keywords from the current query to the user's session.
-     *
-     * @param string $keywords The current query's keywords
-     * @return void
-     */
-    protected function storeKeywordsToSession($keywords)
-    {
-        $currentLastSearches = $GLOBALS['TSFE']->fe_user->getKey(
-            'ses',
-            $this->prefix . '_lastSearches'
-        );
-
-        if (!is_array($currentLastSearches)) {
-            $currentLastSearches = array();
-        }
-
-        $lastSearches = $currentLastSearches;
-        $newLastSearchesCount = array_push($lastSearches, $keywords);
-
-        while ($newLastSearchesCount > $this->configuration->getSearchLastSearchesLimit()) {
-            array_shift($lastSearches);
-            $newLastSearchesCount = count($lastSearches);
-        }
-
-        $GLOBALS['TSFE']->fe_user->setKey(
-            'ses',
-            $this->prefix . '_lastSearches',
-            $lastSearches
-        );
-    }
-
-    /**
-     * Stores the keywords from the current query to the database.
-     *
-     * @param string $keywords The current query's keywords
-     * @return void
-     */
-    protected function storeKeywordsToDatabase($keywords)
-    {
-        $nextSequenceId = $this->getNextSequenceId();
-
-        $GLOBALS['TYPO3_DB']->sql_query(
-            'INSERT INTO tx_solr_last_searches (sequence_id, tstamp, keywords)
-			VALUES ('
-            . $nextSequenceId . ', '
-            . time() . ', '
-            . $GLOBALS['TYPO3_DB']->fullQuoteStr($keywords,
-                'tx_solr_last_searches')
-            . ')
-			ON DUPLICATE KEY UPDATE tstamp = ' . time() . ', keywords = ' . $GLOBALS['TYPO3_DB']->fullQuoteStr($keywords,
-                'tx_solr_last_searches')
-        );
-    }
-
-    /**
-     * Gets the sequence id for the next search entry.
-     *
-     * @return integer The id to be used as the next sequence id for storing the last search keywords.
-     */
-    protected function getNextSequenceId()
-    {
-        $nextSequenceId = 0;
-        $numberOfLastSearchesToLog = $this->configuration['search.']['lastSearches.']['limit'];
-
-        $row = $GLOBALS['TYPO3_DB']->exec_SELECTgetRows(
-            '(sequence_id + 1) % ' . $numberOfLastSearchesToLog . ' as next_sequence_id',
-            'tx_solr_last_searches',
-            '',
-            '',
-            'tstamp DESC',
-            1
-        );
-
-        if (!empty($row)) {
-            $nextSequenceId = $row[0]['next_sequence_id'];
-        }
-
-        return $nextSequenceId;
     }
 }
